Fix error path bugs in fs.putObjectHelper() and internal API bugs in fs.
Code in fs.putObjectHelper() assumes that an error return value from
various fs locking routines means that the returned lock value is nil,
but that is not currently true.

Fix fs.getReadLock()/getWriteLock()/ensureReadLock()/ensureWriteLock()
to return a nil value for the lock if they return an error.

In general, code that calls a "library" routine should not assume
that the other values returned by the library routine are valid when
it returnes an error, but since this assumption is probably not unique
to fs.putObjectHelper() it seems best to return a non-nil value for the
lock only in the case where the lock was acquired.

Change the error code paths in fs.putObjectHelper() that handle locking
failure to set the lock pointer to nil when lock acquisition fails
(previously the lock was not set to nil which would lead to the lock
(which was not acquired) being released in a defer routine).

Fix a couple of cases in fs.putObjectHelper() where errors were ignored
and processing continued (with undefined results) rather then returning
an error to the caller.

Reformat some comments and try to cleanup the code by restructuring it
a bit.

Remove the assumption that if an inode is not a regular file or a symlink
then it must be a directory.  We will be adding more inode types and
that assumption will be false.