Fixed sign-bit error in MountID on C-side of JSON-RPC
This is just a temporary fix as we just avoid setting the high-order bit
that gives the JSON Marshaler on the C-side such fits.

An interim solution is to apply the same "treat it as an int64, not a
uint64, on the C-side as is employed for InodeNumbers that, with the
advent of SnapShots, likely set that bit often.

Ultimately, however, the MountID should be arbitrarily long such that
it may be cryptographically unique more easily (random uint64's are
pretty likely to be unique... but we can do much better easily).